% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e73123_pp}
\alias{e73123_pp}
\title{e73123_pp}
\format{
A data frame with 23520 rows and 6 variables:
\describe{
\item{\code{aaaa}}{a integer denoting the year}
\item{\code{mm}}{a integer denoting the month}
\item{\code{dd}}{a integer denoting the day}
\item{\code{hrs}}{a character denoting the hour}
\item{\code{pp}}{a number of the hourly precipitation}
\item{\code{fecha}}{a date and time of the record}
}
}
\source{
\url{http://imn.ac.cr}
}
\usage{
e73123_pp
}
\description{
Clean meteorological dataset from the station 73123 from the IMN Costa Rica
with precipitation data from 2019-08-31 to 2022-05-19.
}
\keyword{datasets}
