% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utilities.R
\name{quickmerge}
\alias{quickmerge}
\title{To merge big data sets with a lot of rows and an arbitrary number of columns}
\usage{
quickmerge(df1, df2)
}
\arguments{
\item{df1}{first data frame}

\item{df2}{second data frame}
}
\value{
a merge data frame with all the fields from the used data frames
}
\description{
To merge big data sets with a lot of rows and an arbitrary number of columns
}
