% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_utilities.R
\name{ts_format_hr}
\alias{ts_format_hr}
\title{Convert a hourly meteorological table in wide format from the IMN of Costa Rica, to a time series format}
\usage{
ts_format_hr(x, var.type)
}
\arguments{
\item{x}{is a data table from excel or csv}

\item{var.type}{the variable is: pp (precipitation), temp (temperature), hr (relative humidity) or wind}
}
\value{
a data.frame with a date field
}
\description{
Convert a hourly meteorological table in wide format from the IMN of Costa Rica, to a time series format
}
